pyproject.toml
[project]
name = "whittle"
version = "0.1.0"
description = "An AI-powered copilot for OpenFOAM meshing and workflow assistance"
authors = [
    {name = "Kade"}
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "openai>=1.12.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
whittle = "whittle.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100 